bonus_menu: double-precision pulse alpha divide in CalcSelectWait (69.73 -> 69.76)
The target computes the cursor pulse alpha as (double)pulseFrame / 10.0 with a
trailing frsp, not a single-precision fdivs. CalcSelectWait 68.05 -> 68.47.

+++ src/bonus_menu.cpp
@@ -1316,7 +1316,7 @@ void CMenuPcs::CalcSelectWait()
 			}
 			cursor->x = (short)(partySprite->x - 3);
 			cursor->y = (short)(partySprite->y - 8);
-			cursor->alpha = (float)pulseFrame / 10.0f;
+			cursor->alpha = (float)((double)pulseFrame / 10.0);
 		} else {
 			cursor->alpha = 0.0f;
 		}
